+++ Forage/src/main/java/com/forage/repository/TypeDevisRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypeDevisRepository extends JpaRepository<TypeDevis, Long> {
+public interface TypeDevisRepository extends JpaRepository<TypeDevis, Integer> {
 }
